+++ example/windows/flutter/generated_plugins.cmake
@@ -3,6 +3,8 @@
 #
 
 list(APPEND FLUTTER_PLUGIN_LIST
+  agora_rtc_engine
+  permission_handler_windows
 )
 
 set(PLUGIN_BUNDLED_LIBRARIES)