+++ lib/src/layout/grid_layout.dart
@@ -16,18 +16,18 @@ class GridLayout extends StatefulWidget {
   final Widget? disabledVideoWidget;
 
   /// Render mode for local and remote video
-  final VideoRenderMode? videoRenderMode;
+  final VideoRenderMode videoRenderMode;
 
   const GridLayout({
     Key? key,
     required this.client,
     this.showNumberOfUsers,
     this.disabledVideoWidget = const DisabledVideoWidget(),
-    this.videoRenderMode,
+    this.videoRenderMode = VideoRenderMode.Hidden,
   }) : super(key: key);
 
   @override
-  _GridLayoutState createState() => _GridLayoutState();
+  State<GridLayout> createState() => _GridLayoutState();
 }
 
 class _GridLayoutState extends State<GridLayout> {
@@ -169,7 +169,7 @@ class DisabledVideoStfWidget extends StatefulWidget {
       : super(key: key);
 
   @override
-  _DisabledVideoStfWidgetState createState() => _DisabledVideoStfWidgetState();
+  State<DisabledVideoStfWidget> createState() => _DisabledVideoStfWidgetState();
 }
 
 class _DisabledVideoStfWidgetState extends State<DisabledVideoStfWidget> {